enable support for custom session configurations
